code cleanup after LINQPad 4 & 5 verification
